remove Renderer interface
