--- include/dirkengine.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "core/globals.hpp"
-#include "engine/actor.hpp"
-#include "engine/dirkengine.hpp"
-#include "engine/game_instance.hpp"
-#include "render/render_types.hpp"
-#include "render/renderer.hpp"
-#include "render/renderer_types.hpp"
-#include "resources/resource_manager.hpp"
